selector data from single to dual/multipe
